requirements.txt
streamlit>=1.40.0
supabase>=2.0.0
requests>=2.31.0
pandas>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0
plotly>=5.18.0
python-dotenv>=1.0.0
